refactor: composables でリファクタ
無限呼び出しがされることと、初期選択がうまくされない不具合あり
